requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==42.0.5
distro==1.9.0
dnspython==2.8.0
email_validator==2.1.1
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.3
google-api-core==2.28.1
google-api-python-client==2.94.0
google-auth==2.21.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.72.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.6
jiter==0.12.0
keyring==23.13.0
keyrings.alt==5.0.2
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
msal==1.23.0
oauthlib==3.3.1
openai==2.8.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
Werkzeug==2.3.7
